python-tests/veo/requirements.txt
# Environment variables
python-dotenv==1.0.0

google-genai>1.8.0